claire: commit changement id Long Admin

+++ Projet_ECommerce/src/fr/adaming/model/Admin.java
@@ -19,7 +19,7 @@ public class Admin implements Serializable{
 	@Id
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
 	@Column(name="id_admin")
-	private int id;
+	private Long id;
 	private String login;
 	private String password;
 	
@@ -44,7 +44,7 @@ public class Admin implements Serializable{
 	 * @param login
 	 * @param password
 	 */
-	public Admin(int id, String login, String password) {
+	public Admin(Long id, String login, String password) {
 		super();
 		this.id = id;
 		this.login = login;
@@ -55,13 +55,13 @@ public class Admin implements Serializable{
 	/**
 	 * @return the id
 	 */
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
 	/**
 	 * @param id the id to set
 	 */
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 	/**
